+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 08-04-2025
 ### Added
 - More functionality to Debugger.cs for printing
+- TCPClient.cs and partial functionality
+- ITrasnportClient.cs as an interface for TCP and UDP client variations
 ### Updated
 - .gitignore
+- Makefile
 ### Changed
 - Program.cs to accompany future implementations